fix syncLogDequeSerializer bug & use FixedThreadPool to replace the ScheduledThreadPoolExecutor when check the log applied

+++ cluster/src/main/java/org/apache/iotdb/cluster/log/manage/serializable/SyncLogDequeSerializer.java
@@ -542,8 +542,8 @@ public class SyncLogDequeSerializer implements StableEntryManager {
     if (!success) {
       logger.error("recover log index file failed, clear all logs in disk, {}",
           lastIndexFile.getAbsoluteFile());
-      for (int i = 0; i < logIndexFileList.size(); i++) {
-        deleteLogDataAndIndexFile(i);
+      while (!logIndexFileList.isEmpty()) {
+        deleteTheFirstLogDataAndIndexFile();
       }
       clearFirstLogIndex();
 
@@ -561,8 +561,8 @@ public class SyncLogDequeSerializer implements StableEntryManager {
     if (!success) {
       logger.error("recover log data file failed, clear all logs in disk,{}",
           lastDataFile.getAbsoluteFile());
-      for (int i = 0; i < logIndexFileList.size(); i++) {
-        deleteLogDataAndIndexFile(i);
+      while (!logIndexFileList.isEmpty()) {
+        deleteTheFirstLogDataAndIndexFile();
       }
       clearFirstLogIndex();
     }
@@ -837,8 +837,8 @@ public class SyncLogDequeSerializer implements StableEntryManager {
       // 1. delete
       forceFlushLogBuffer();
       closeCurrentFile(meta.getCommitLogIndex());
-      for (int i = 0; i < logDataFileList.size(); i++) {
-        deleteLogDataAndIndexFile(i);
+      while (!logDataFileList.isEmpty()) {
+        deleteTheFirstLogDataAndIndexFile();
       }
       deleteMetaFile();
 
@@ -891,8 +891,8 @@ public class SyncLogDequeSerializer implements StableEntryManager {
 
   public void checkDeletePersistRaftLog() {
     // 1. check the log index offset list size
+    lock.lock();
     try {
-      lock.lock();
       if (logIndexOffsetList.size() > maxRaftLogIndexSizeInMemory) {
         int compactIndex = logIndexOffsetList.size() - maxRaftLogIndexSizeInMemory;
         logIndexOffsetList.subList(0, compactIndex).clear();
@@ -906,7 +906,7 @@ public class SyncLogDequeSerializer implements StableEntryManager {
     lock.lock();
     try {
       while (logDataFileList.size() > maxNumberOfPersistRaftLogFiles) {
-        deleteLogDataAndIndexFile(0);
+        deleteTheFirstLogDataAndIndexFile();
       }
     } finally {
       lock.unlock();
@@ -919,7 +919,7 @@ public class SyncLogDequeSerializer implements StableEntryManager {
         File firstFile = logDataFileList.get(0);
         String[] splits = firstFile.getName().split(FILE_NAME_SEPARATOR);
         if (meta.getCommitLogIndex() - Long.parseLong(splits[1]) > maxPersistRaftLogNumberOnDisk) {
-          deleteLogDataAndIndexFile(0);
+          deleteTheFirstLogDataAndIndexFile();
         } else {
           return;
         }
@@ -927,26 +927,32 @@ public class SyncLogDequeSerializer implements StableEntryManager {
     } finally {
       lock.unlock();
     }
-
   }
 
-  private void deleteLogDataAndIndexFile(int index) {
+  private void deleteTheFirstLogDataAndIndexFile() {
+    if (logDataFileList.isEmpty()) {
+      return;
+    }
     File logDataFile = null;
     File logIndexFile = null;
+
+    lock.lock();
     try {
-      lock.lock();
-      logDataFile = logDataFileList.get(index);
-      logIndexFile = logIndexFileList.get(index);
+      logDataFile = logDataFileList.get(0);
+      logIndexFile = logIndexFileList.get(0);
+      if (logDataFile == null || logIndexFile == null) {
+        logger.error("the log data or index file is null, some error occurred");
+        return;
+      }
       Files.delete(logDataFile.toPath());
       Files.delete(logIndexFile.toPath());
-      logDataFileList.remove(index);
-      logIndexFileList.remove(index);
+      logDataFileList.remove(0);
+      logIndexFileList.remove(0);
       logger.debug("delete date file={}, index file={}", logDataFile.getAbsoluteFile(),
           logIndexFile.getAbsoluteFile());
     } catch (IOException e) {
-      logger.error("delete file failed, index={}, data file={}, index file={}", index,
-          logDataFile == null ? null : logDataFile.getAbsoluteFile(),
-          logIndexFile == null ? null : logIndexFile.getAbsoluteFile());
+      logger.error("delete file failed, data file={}, index file={}",
+          logDataFile.getAbsoluteFile(), logIndexFile.getAbsoluteFile());
     } finally {
       lock.unlock();
     }
